requirements_core.txt
dj-database-url==0.3.0
django-appconf==0.6
django-classy-tags==0.5.1
django-cms==3.0
django-compressor==1.3
django-filter==0.7
django-hvad==0.3
django-mptt==0.6.0
django-reversion==1.8.0
django-taggit==0.12
django-wysiwyg-redactor==0.3.9
Django==1.5.5
djangocms-admin-style==0.2.2
djangocms-column==1.3
djangocms-file==0.0.1
djangocms-flash==0.0.2
djangocms-googlemap==0.0.5
djangocms-inherit==0.0.1
djangocms-installer==0.4.1
djangocms-link==1.3.4
djangocms-picture==0.0.2
djangocms-style==1.3
djangocms-teaser==0.0.1
djangocms-text-ckeditor==2.1.4
djangocms-video==0.0.1
djangorestframework==2.3.13
html5lib==0.999
Markdown==2.4
Pillow==2.4.0
pytz==2014.2
six==1.6.1
South==0.8.4